Fixing minor bugs (listed in description)
validation is not working
camera is rotating with tank when input field is selected. it sould just follow it depending on followPlayer value
log message on script save

is it saving and loading scripts using the same location?
yes. Except when you do Ctrl+B in the editor - it saves files in the project folder.